use generalized wrapper for console apps
